Split pane manager helpers
