+++ src/renderer/src/lib/pane-manager/pane-lifecycle.ts
@@ -22,41 +22,13 @@ import {
   attachPaneFitResizeObserver,
   detachPaneFitResizeObserver
 } from './pane-fit-resize-observer'
+import { buildDefaultTerminalOptions } from './pane-terminal-options'
 
 // ---------------------------------------------------------------------------
 // Pane creation, terminal open/close, addon management
 // ---------------------------------------------------------------------------
 
 const ENABLE_WEBGL_RENDERER = true
-
-export function buildDefaultTerminalOptions(): ITerminalOptions {
-  return {
-    allowProposedApi: true,
-    cursorBlink: true,
-    cursorStyle: 'bar',
-    fontSize: 14,
-    // Cross-platform fallback chain — ensures the terminal can always find a
-    // usable monospace font regardless of OS, even if user settings haven't
-    // loaded yet. macOS-only fonts are harmlessly skipped on other platforms.
-    // Must stay in sync with FALLBACK_FONTS in layout-serialization.ts; the
-    // trailing Nerd Fonts let Powerline/PUA glyphs render even at first paint
-    // before the user's configured terminalFontFamily is applied.
-    fontFamily:
-      '"SF Mono", "Menlo", "Monaco", "Cascadia Mono", "Consolas", "DejaVu Sans Mono", "Liberation Mono", "Symbols Nerd Font Mono", "MesloLGS Nerd Font", "JetBrainsMono Nerd Font", "Hack Nerd Font", monospace',
-    fontWeight: '300',
-    fontWeightBold: '500',
-    scrollback: 10000,
-    allowTransparency: false,
-    // Why: on macOS, non-US layouts rely on Option to compose real characters
-    // like @ (German Option+L) and € (German Option+E). Enabling xterm's
-    // Meta mode here makes Option behave like Esc+key instead, which steals
-    // those composed characters before they reach the shell.
-    // Readline shortcuts (Option+B/F/D) are compensated in terminal-shortcut-policy.ts.
-    macOptionIsMeta: false,
-    macOptionClickForcesSelection: true,
-    drawBoldTextInBrightColors: true
-  }
-}
 
 function getTerminalUrlOpenHint(): string {
   return navigator.userAgent.includes('Mac')
